Removed circular dependencies
