+++ include/xtensor/xreducer.hpp
@@ -31,7 +31,6 @@
 #include "xexpression.hpp"
 #include "xgenerator.hpp"
 #include "xiterable.hpp"
-#include "xreducer.hpp"
 #include "xtensor_config.hpp"
 #include "xutils.hpp"
